// src/college/college-info/dto/college-grouped-response.dto.ts
export class CollegeGroupedResponseDto {
  college_information: {
    college_id?: number;
    created_at?: Date;
    updated_at?: Date;
    is_active?: boolean | true;
    college_name?: string;
    short_name?: string;
    search_names?: string;
    parent_college_id?: number;
    city_id?: number;
    state_id?: number;
    country_id?: number;
    location?: string;
    PIN_code?: string;
    latitude_longitude?: string;
    college_email?: string;
    college_phone?: string;
    college_website?: string;
    type_of_institute?: string;
    affiliated_university_id?: number;
    founded_year?: string;
    logo_img?: string;
    banner_img?: string;
    total_student?: number;
    campus_size?: number;
    UGC_approved?: boolean;
    kapp_rating?: number;
    kapp_score?: number;
    city?: string | null;
    state?: string | null;
    country?: string | null;
    primary_stream_id?: number | null;
    nacc_grade?: string | null;
    slug?: string;
    girls_only?: boolean;
    is_university?: boolean;
    course_count?: number;
    meta_desc?: string;
    is_online?: boolean;
    college_brochure?: string;
  };
  news_section?: CollegeContentDto[];
  info_section?: CollegeContentDto[];
  popular_courses?: any[];
  exam_section?: ExamSectionDto[];
}

export class CollegeGroupedHighlightsResponseDto {
  college_information: {
    college_id?: number;
    created_at?: Date;
    updated_at?: Date;
    is_active?: boolean | true;
    college_name?: string;
    short_name?: string;
    search_names?: string;
    parent_college_id?: number;
    city_id?: number;
    state_id?: number;
    country_id?: number;
    location?: string;
    PIN_code?: string;
    latitude_longitude?: string;
    college_email?: string;
    college_phone?: string;
    college_website?: string;
    type_of_institute?: string;
    affiliated_university_id?: number;
    founded_year?: string;
    logo_img?: string;
    banner_img?: string;
    total_student?: number;
    campus_size?: number;
    UGC_approved?: boolean;
    kapp_rating?: number;
    kapp_score?: number;
    city?: string | null;
    state?: string | null;
    country?: string | null;
    primary_stream_id?: number | null;
    nacc_grade?: string | null;
    slug?: string;
    girls_only?: boolean;
    is_university?: boolean;
    course_count?: number;
    meta_desc?: string;
    is_online?: boolean;
    college_brochure?: string;
  };
  news_section?: CollegeContentDto[];
  highlight_section?: CollegeContentDto[];
  exam_section?: ExamSectionDto[];
}

export class CollegeGroupedScholarshipResponseDto {
  college_information: {
    college_id?: number;
    created_at?: Date;
    updated_at?: Date;
    is_active?: boolean | true;
    college_name?: string;
    short_name?: string;
    search_names?: string;
    parent_college_id?: number;
    city_id?: number;
    state_id?: number;
    country_id?: number;
    location?: string;
    PIN_code?: string;
    latitude_longitude?: string;
    college_email?: string;
    college_phone?: string;
    college_website?: string;
    type_of_institute?: string;
    affiliated_university_id?: number;
    founded_year?: string;
    logo_img?: string;
    banner_img?: string;
    total_student?: number;
    campus_size?: number;
    UGC_approved?: boolean;
    kapp_rating?: number;
    kapp_score?: number;
    city?: string | null;
    state?: string | null;
    country?: string | null;
    primary_stream_id?: number | null;
    nacc_grade?: string | null;
    slug?: string;
    girls_only?: boolean;
    is_university?: boolean;
    course_count?: number;
    meta_desc?: string;
    is_online?: boolean;
    college_brochure?: string;
  };
  news_section?: CollegeContentDto[];
  scholarship_section?: {
    description?: string;
  }[];
  exam_section?: ExamSectionDto[];
}

export class CollegeNewsResponseDto {
  college_information: {
    college_id?: number;
    created_at?: Date;
    updated_at?: Date;
    is_active?: boolean | true;
    college_name?: string;
    short_name?: string;
    search_names?: string;
    parent_college_id?: number;
    city_id?: number;
    state_id?: number;
    country_id?: number;
    location?: string;
    PIN_code?: string;
    latitude_longitude?: string;
    college_email?: string;
    college_phone?: string;
    college_website?: string;
    type_of_institute?: string;
    affiliated_university_id?: number;
    founded_year?: string;
    logo_img?: string;
    banner_img?: string;
    total_student?: number;
    campus_size?: number;
    UGC_approved?: boolean;
    kapp_rating?: number;
    kapp_score?: number;
    city?: string | null;
    state?: string | null;
    country?: string | null;
    primary_stream_id?: number | null;
    nacc_grade?: string | null;
    slug?: string;
    girls_only?: boolean;
    is_university?: boolean;
    course_count?: number;
    meta_desc?: string;
    is_online?: boolean;
    college_brochure?: string;
  };
  news_section?: CollegeContentDto[];
}

export class CollegeWiseNewsResponseDto {
  college_information: {
    college_id?: number;
    created_at?: Date;
    updated_at?: Date;
    is_active?: boolean | true;
    college_name?: string;
    short_name?: string;
    search_names?: string;
    parent_college_id?: number;
    city_id?: number;
    state_id?: number;
    country_id?: number;
    location?: string;
    PIN_code?: string;
    latitude_longitude?: string;
    college_email?: string;
    college_phone?: string;
    college_website?: string;
    type_of_institute?: string;
    affiliated_university_id?: number;
    founded_year?: string;
    logo_img?: string;
    banner_img?: string;
    total_student?: number;
    campus_size?: number;
    UGC_approved?: boolean;
    kapp_rating?: number;
    kapp_score?: number;
    city?: string | null;
    state?: string | null;
    country?: string | null;
    primary_stream_id?: number | null;
    nacc_grade?: string | null;
    slug?: string;
    girls_only?: boolean;
    is_university?: boolean;
    course_count?: number;
    meta_desc?: string;
    is_online?: boolean;
    college_brochure?: string;
  };
  news_section?: {
    college_content_id?: number;
    silos?: string;
    title?: string;
    description?: string;
    updated_at?: Date;
    is_active?: boolean;
    author_name?: string;
    author_img?: string;
    author_id?: number;
  };
}